Fix warnings

+++ src/Core/Algorithms/Math/Tests/AppendMatrixTests.cc
@@ -41,8 +41,8 @@ namespace
   DenseMatrixHandle matrix1()
   {
     DenseMatrixHandle m(boost::make_shared<DenseMatrix>(3, 4));
-    for (size_t i = 0; i < m->rows(); ++ i)
-      for (size_t j = 0; j < m->cols(); ++ j)
+    for (int i = 0; i < m->rows(); ++ i)
+      for (int j = 0; j < m->cols(); ++ j)
         (*m)(i, j) = 3.0 * i + j - 5;
     return m;
   }
